# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/adminlogin": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Authenticate user",
                "operationId": "b6706dcb4c0805d1dd0194d52e958280",
                "parameters": [
                    {
                        "name": "userName",
                        "in": "query",
                        "description": "User's username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/get_amenities/{sortOrder}": {
            "get": {
                "tags": [
                    "Amenities"
                ],
                "summary": "Amenities list",
                "operationId": "ba989bcfbb96c5904c789409ee37e4c4",
                "parameters": [
                    {
                        "name": "sortOrder",
                        "in": "query",
                        "description": "Amenities id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenities List"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                }
            }
        },
        "/api/auth/store_amenities": {
            "post": {
                "tags": [
                    "Amenities"
                ],
                "summary": "Store Amenities",
                "operationId": "5d8660a071dc99f5c2314d0eb4f8ab96",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Amenities name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "description": "Amenities URL",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Amenities business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenities Created"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/update_amenities/{id}": {
            "post": {
                "tags": [
                    "Amenities"
                ],
                "summary": "Update Amenities",
                "operationId": "da681f74d4f02317dc9a01a7eca0fe1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Amenities id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Amenities name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "url",
                        "in": "query",
                        "description": "Amenities url",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "business_key",
                        "in": "query",
                        "description": "Amenities business_key",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Amenities Updated"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/delete_amenities/{id}": {
            "get": {
                "tags": [
                    "Amenities"
                ],
                "summary": "Delete Amenities",
                "operationId": "fc767e39f94aa2f9a3bc849fe330cfe1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Amenities id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Primary Service"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}